Convert dictIterator to use 64 bit sizes
Also converts dict to use a 64-bit type instead
of a compiler-provided "long" size.

Also converts return type for retrieving a dict
key index (_dictKeyIndex) to 64 bits as well.

Attempt to fix issue #1814

+++ src/dict.c
@@ -62,7 +62,7 @@ static unsigned int dict_force_resize_ratio = 5;
 
 static int _dictExpandIfNeeded(dict *ht);
 static unsigned long _dictNextPower(unsigned long size);
-static int _dictKeyIndex(dict *ht, const void *key);
+static uint64_t _dictKeyIndex(dict *ht, const void *key);
 static int _dictInit(dict *ht, dictType *type, void *privDataPtr);
 
 /* -------------------------- hash functions -------------------------------- */
@@ -337,7 +337,7 @@ int dictAdd(dict *d, void *key, void *val)
  */
 dictEntry *dictAddRaw(dict *d, void *key)
 {
-    int index;
+    uint64_t index;
     dictEntry *entry;
     dictht *ht;
 
@@ -576,7 +576,7 @@ dictEntry *dictNext(dictIterator *iter)
                     iter->fingerprint = dictFingerprint(iter->d);
             }
             iter->index++;
-            if (iter->index >= (signed) ht->size) {
+            if (iter->index >= ht->size) {
                 if (dictIsRehashing(iter->d) && iter->table == 0) {
                     iter->table++;
                     iter->index = 0;
@@ -911,9 +911,9 @@ static unsigned long _dictNextPower(unsigned long size)
  *
  * Note that if we are in the process of rehashing the hash table, the
  * index is always returned in the context of the second (new) hash table. */
-static int _dictKeyIndex(dict *d, const void *key)
+static uint64_t _dictKeyIndex(dict *d, const void *key)
 {
-    unsigned int h, idx, table;
+    uint64_t h, idx, table;
     dictEntry *he;
 
     /* Expand the hash table if needed */
